refactor: remove deprecated configuration command functions and enhance package documentation
